% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{add.genexp}
\alias{add.genexp}
\title{Method to add a slot of expression to \code{MultiDataSet}.}
\usage{
add.genexp(object, gexpSet, warnings = TRUE)
}
\arguments{
\item{object}{\code{MultiDataSet} that will be filled.}

\item{gexpSet}{\code{ExpressionSet} to be used to fill the slot.}

\item{warnings}{Logical to indicate if warnings will be displayed.}
}
\value{
A new \code{MultiDataSet} with the slot \code{"expression"}
filled.
}
\description{
This method adds or overwrites the slot \code{"expression"} of an
\code{MultiDataSet} with the content of the given \code{ExpressionSet}.
}
\examples{
multi <- new("MultiDataSet")
eset <- new("ExpressionSet", exprs = matrix(runif(4), 2))
fData(eset) <- data.frame(chromosome = c("chr1", "chr2"), start = c(12414, 1234321),
 end = c(121241, 12412412414), stringsAsFactors = FALSE)
multi <- add.genexp(multi, eset)
}

