% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{add.methy}
\alias{add.methy}
\alias{add.methy-methods}
\title{Method to add a slot of methylation to \code{MultiDataSet}.}
\usage{
add.methy(object, methySet, warnings = TRUE)
}
\arguments{
\item{object}{\code{MultiDataSet} that will be filled.}

\item{methySet}{\code{MethylationSet} to be used to fill the slot.}

\item{warnings}{Logical to indicate if warnings will be displayed.}
}
\value{
A new \code{MultiDataSet} with the slot \code{"methylation"}
filled.
}
\description{
This method adds or overwrites the slot \code{"methylation"} of an
\code{MultiDataSet} with the content of the given \code{MethylationSet}.
}
\examples{
if (require(MEALData)){
 multi <- new("MultiDataSet")
 betavals <- betavals[1:100, ]  ## To speed up the example, the beta values are reduced
 methy <- prepareMethylationSet(betavals, pheno)
 multi <- add.methy(multi, methy)
}
}

