% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{add.set}
\alias{add.set}
\alias{add.set-methods}
\title{Method to add a slot to \code{MultiDataSet}.}
\usage{
add.set(object, set, dataset.name, warnings = TRUE)
}
\arguments{
\item{object}{\code{MultiDataSet} that will be filled.}

\item{set}{Object derived from \code{eSet} to be used to fill the slot.}

\item{dataset.name}{Character with the name of the slot to be filled.}

\item{warnings}{Logical to indicate if warnings will be displayed.}
}
\value{
A new \code{MultiDataSet} with a slot filled.
}
\description{
This method adds or overwrites a slot of a \code{MultiDataSet} with the content
of the given \code{eSet}.
}
\examples{
multi <- new("MultiDataSet")
eset <- new("ExpressionSet", exprs = matrix(runif(10), 5))
multi <- add.set(multi, eset, "exampledata")
}

