% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{add.snps}
\alias{add.snps}
\title{Method to add a slot of SNPs to \code{MultiDataSet}.}
\usage{
add.snps(object, snpSet, warnings = TRUE)
}
\arguments{
\item{object}{\code{MultiDataSet} that will be filled.}

\item{snpSet}{\code{SnpSet} to be used to fill the slot.}

\item{warnings}{Logical to indicate if warnings will be displayed.}
}
\value{
A new \code{MultiDataSet} with the slot \code{"snps"}
filled.
}
\description{
This method adds or overwrites the slot \code{"snps"} of an
\code{MultiDataSet} with the content of the given \code{SnpSet}.
}

