% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AllClasses.R,
%   R/AnalysisRegionResults_Constructor.R,
%   R/AnalysisRegionResults_SettersGetters.R, R/AnalysisRegionResults_plotRDA.R,
%   R/AnalysisRegionResults_topRDAhits.R
\docType{class}
\name{AnalysisRegionResults}
\alias{AnalysisRegionResults}
\alias{AnalysisRegionResults-class}
\alias{AnalysisRegionResults-methods}
\alias{analysisRegionResults}
\alias{AnalysisRegionResults-methods}
\alias{getRange,AnalysisRegionResults-method}
\alias{AnalysisRegionResults-methods}
\alias{getRange}
\alias{getRDA,AnalysisRegionResults-method}
\alias{AnalysisRegionResults-methods}
\alias{getRDA}
\alias{globalPval,AnalysisRegionResults-method}
\alias{AnalysisRegionResults-methods}
\alias{globalPval}
\alias{globalR2,AnalysisRegionResults-method}
\alias{AnalysisRegionResults-methods}
\alias{globalR2}
\alias{RDAPval,AnalysisRegionResults-method}
\alias{AnalysisRegionResults-methods}
\alias{RDAPval}
\alias{regionR2,AnalysisRegionResults-method}
\alias{AnalysisRegionResults-methods}
\alias{regionR2}
\alias{plotRDA,AnalysisRegionResults-method}
\alias{AnalysisRegionResults-methods}
\alias{topRDAhits,AnalysisRegionResults-method}
\alias{AnalysisRegionResults-methods}
\title{AnalysisRegionResults instances}
\usage{
analysisRegionResults(analysisResults, range, rdaRes)

\S4method{getRange}{AnalysisRegionResults}(object)

\S4method{getRDA}{AnalysisRegionResults}(object)

\S4method{globalPval}{AnalysisRegionResults}(object)

\S4method{globalR2}{AnalysisRegionResults}(object)

\S4method{RDAPval}{AnalysisRegionResults}(object)

\S4method{regionR2}{AnalysisRegionResults}(object)

\S4method{plotRDA}{AnalysisRegionResults}(object, n_feat = 5,
  main = "RDA plot")

\S4method{topRDAhits}{AnalysisRegionResults}(object, pval = 0.05)
}
\arguments{
\item{analysisResults}{\code{AnalysisResults}}

\item{range}{\code{GenomicRanges}}

\item{rdaRes}{List with RDA results}

\item{object}{\code{MethylationResults}}

\item{n_feat}{Numeric with the number of features to be highlighted.}

\item{main}{Character with the plot title.}

\item{pval}{numeric with the p-value threshold. Only features with a p-values below this threshold will be shown.}
}
\value{
An \code{AnalysisRegionResults}
}
\description{
\code{AnalysisResults} heir with the analyses performed in a range of the 
whole genome.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getRange}: Get range where the analyses was performed

\item \code{getRDA}: Get rda object.

\item \code{globalPval}: Get global p-value.

\item \code{globalR2}: Get global R2.

\item \code{RDAPval}: Get p-value of RDA.

\item \code{regionR2}: Get R2 of the RDA model

\item \code{plotRDA}: Plot RDA results

\item \code{topRDAhits}: Get the top features associated with the RDA
}}

\section{Slots}{

\describe{
\item{\code{range}}{\code{GenomicRanges} used to perform the analysis.}

\item{\code{rda}}{\code{rda} object from \code{vegan} package with the results of RDA
analysis in the range.}

\item{\code{regionR2}}{Numeric with the R2 of the region calculated using a redundancy analysis.}

\item{\code{RDAPval}}{Numeric with the p-value of the RDA.}

\item{\code{globalR2}}{Numeric with the global R2.}

\item{\code{globalPval}}{Numeric with the probability of finding a region with the same number of probes with
a bigger R2.}
}}

\examples{
showClass("AnalysisRegionResults")
}
