% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisRegionResults_Constructor.R
\name{analysisRegionResults}
\alias{analysisRegionResults}
\title{Old class to encapsulate results (deprecated in new version)}
\usage{
analysisRegionResults()
}
\value{
Deprecated
}
\description{
Old class to encapsulate results (deprecated in new version)
}
\examples{
analysisRegionResults()
}
