% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_AllClasses.R, R/AnalysisResults_Constructor.R, R/AnalysisResults_SettersGetters.R, R/AnalysisResults_exportResults.R, R/AnalysisResults_plotEWAS.R, R/AnalysisResults_plotQQ.R, R/AnalysisResults_plotRegion.R, R/AnalysisResults_plotVolcano.R
\docType{class}
\name{AnalysisResults}
\alias{AnalysisResults}
\alias{AnalysisResults-class}
\alias{AnalysisResults-methods}
\alias{analysisResults}
\alias{blocks}
\alias{blocks,AnalysisResults-method}
\alias{bumps}
\alias{bumps,AnalysisResults-method}
\alias{covariableNames}
\alias{covariableNames,AnalysisResults-method}
\alias{dmrCate}
\alias{dmrCate,AnalysisResults-method}
\alias{exportResults,AnalysisResults-method}
\alias{feats}
\alias{feats,AnalysisResults-method}
\alias{featvals}
\alias{featvals,AnalysisResults-method}
\alias{getGeneVals,AnalysisResults-method}
\alias{getMs,AnalysisResults-method}
\alias{model}
\alias{model,AnalysisResults-method}
\alias{modelVariables}
\alias{modelVariables,AnalysisResults-method}
\alias{pData,AnalysisResults-method}
\alias{pData<-,AnalysisResults,ANY-method}
\alias{phenoData,AnalysisResults-method}
\alias{phenoData<-,AnalysisResults,ANY-method}
\alias{plotEWAS,AnalysisResults-method}
\alias{plotQQ,AnalysisResults-method}
\alias{plotRegion,AnalysisResults-method}
\alias{plotVolcano,AnalysisResults-method}
\alias{probeResults}
\alias{probeResults,AnalysisResults-method}
\alias{regionResults}
\alias{regionResults,AnalysisResults-method}
\alias{sampleNames,AnalysisResults-method}
\alias{variableNames}
\alias{variableNames,AnalysisResults-method}
\title{AnalysisResults instances}
\usage{
analysisResults(set, model, regionResults, probeResults, num_feat = 50,
  num_vars = ncol(pData(set)))

\S4method{blocks}{AnalysisResults}(object)

\S4method{bumps}{AnalysisResults}(object)

\S4method{covariableNames}{AnalysisResults}(object)

\S4method{dmrCate}{AnalysisResults}(object)

\S4method{feats}{AnalysisResults}(object)

\S4method{featvals}{AnalysisResults}(object)

\S4method{getGeneVals}{AnalysisResults}(object, gene)

\S4method{getMs}{AnalysisResults}(object, threshold = 1e-04)

\S4method{model}{AnalysisResults}(object)

\S4method{modelVariables}{AnalysisResults}(object)

\S4method{phenoData}{AnalysisResults}(object)

\S4method{phenoData}{AnalysisResults,ANY}(object) <- value

\S4method{pData}{AnalysisResults}(object)

\S4method{pData}{AnalysisResults,ANY}(object) <- value

\S4method{probeResults}{AnalysisResults}(object)

\S4method{regionResults}{AnalysisResults}(object)

\S4method{sampleNames}{AnalysisResults}(object)

\S4method{variableNames}{AnalysisResults}(object)

\S4method{exportResults}{AnalysisResults}(object, dir = "./", prefix = NULL,
  vars = modelVariables(object))

\S4method{plotEWAS}{AnalysisResults}(object,
  variable = modelVariables(object)[1], range = NULL)

\S4method{plotQQ}{AnalysisResults}(object,
  variable = modelVariables(object)[1])

\S4method{plotRegion}{AnalysisResults}(object,
  variable = modelVariables(object)[[1]], range = NULL)

\S4method{plotVolcano}{AnalysisResults}(object,
  variable = modelVariables(object)[1])
}
\arguments{
\item{set}{\code{MethylationSet} or \code{ExpressionSet} used to perform the analysis}

\item{model}{Model matrix used to produce the calculations}

\item{regionResults}{List with the region results}

\item{probeResults}{List with the probe results}

\item{num_feat}{Numeric with the minimum number of feature values to be included.}

\item{num_vars}{Numeric with the number of columns of the pData table
that should be considered as variables.}

\item{object}{\code{AnalysisResults}}

\item{gene}{Character with the name of the gene}

\item{threshold}{Numeric with the threshold to avoid 0s and 1s.}

\item{value}{AnnotatedDataFrame or data.frame with the phenotype}

\item{dir}{Character with the path to export.}

\item{prefix}{Character with a prefix to be added to all file names.}

\item{vars}{Character vector with the names of the variables to be exported. Note:
names should be that of the model.}

\item{variable}{Character with the variable name used to obtain the probe results.
Note: model name should be used. Original variable name might not be valid.}

\item{range}{GenomicRange whose probes will be highlighted}
}
\value{
\code{AnalysisResults}
}
\description{
Container with the results of per probe and per region analyses.

AnalysisResults instances
}
\section{Methods (by generic)}{
\itemize{
\item \code{blocks}: Get BlockFinder analysis results

\item \code{bumps}: Get Bumphunter analysis results

\item \code{covariableNames}: Get covariable names

\item \code{dmrCate}: Get dmrCate analysis results

\item \code{feats}: Get features names

\item \code{featvals}: Get features values matrix

\item \code{getGeneVals}: Get probe results of a gene

\item \code{getMs}: Get Ms values

\item \code{model}: Get model used to perform the analysis

\item \code{modelVariables}: Get names of the variables in the model matrix

\item \code{phenoData}: Get phenotypes data (AnnotatedDataFrame)

\item \code{phenoData<-}: Set phenotypes data (AnnotatedDataFrame)

\item \code{pData}: Get phenotypes data (data.frame)

\item \code{pData<-}: Set phenotypes data (data.frame)

\item \code{probeResults}: Get per probe analysis results

\item \code{regionResults}: Get all per region analysis results

\item \code{sampleNames}: Get sample names

\item \code{variableNames}: Get variable names

\item \code{exportResults}: Exports results data.frames to csv files.

\item \code{plotEWAS}: Plot a Manhattan plot with the probe results

\item \code{plotQQ}: QQ plot of probe analysis

\item \code{plotRegion}: Plot of the region

\item \code{plotVolcano}: Make a Volcano plot with the probe results
}}
\section{Slots}{

\describe{
\item{\code{originalclass}}{Character with the class of the object used to perform the
analysis}

\item{\code{features}}{Matrix with the values of the most significant features.}

\item{\code{phenotypes}}{AnnotatedDataFrame with the phenotypes.}

\item{\code{model}}{Matrix with the model used in the analysis}

\item{\code{sampleNames}}{Character vector with the names of the samples}

\item{\code{variableNames}}{Character vector with the names of the variables used in the
analysis. Names are equal to those find in phenotypes.}

\item{\code{covariableNames}}{Character vector with the names of the covariables used in the
analysis. Names are equal to those find in phenotypes.}

\item{\code{results}}{List of data.frames with the results of per probe analysis. Names are
those of the model.}

\item{\code{DMRcate}}{List of data.frames with the results of \code{DMRcate}. Names are
those of the model.}

\item{\code{Bumphunter}}{List of data.frames with the results of \code{Bumphunter}. Names are
those of the model.}

\item{\code{BlockFinder}}{List of data.frames with the results of \code{BlockFinder}. Names are
those of the model.}
}}
\examples{
showClass("AnalysisResults")
}

