% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisResults_Constructor.R
\name{analysisResults}
\alias{analysisResults}
\title{Old class to encapsulate results (deprecated in new version)}
\usage{
analysisResults()
}
\value{
Deprecated
}
\description{
Old class to encapsulate results (deprecated in new version)
}
\examples{
analysisResults()
}
