% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRelevantSNPs.R
\name{calculateRelevantSNPs}
\alias{calculateRelevantSNPs}
\title{Calculate the SNPs correlated to cpgs}
\usage{
calculateRelevantSNPs()
}
\description{
This function is provided for compatibility with older versions
only, and will be defunct at the next release.
}
\examples{
calculateRelevantSNPs()
}
