% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{checkProbes}
\alias{checkProbes}
\title{Filter \code{MethylationSet} probes}
\usage{
checkProbes(object)
}
\arguments{
\item{object}{\code{MethylationSet}}
}
\value{
\code{MethylationSet} containing the common samples.
}
\description{
This function selects probes present in the annotation matrix. Probes without
annotation and annotation values without beta values are discarded.
}
\examples{
if (require(MEALData)){
 betavals <- betavals[1:100, ]  ## To speed up the example, the beta values are reduced
 methy <- prepareMethylationSet(betavals, pheno)
 checkProbes(methy)
}
}

