% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{checkSamples}
\alias{checkSamples}
\title{Modify a \code{MethylationSet} to only contain common samples}
\usage{
checkSamples(object)
}
\arguments{
\item{object}{\code{MethylationSet}}
}
\value{
\code{MethylationSet} containing the common samples.
}
\description{
This function removes samples that have beta values but no phenotypes and vice versa.
If snps object is present, only samples present in the three set are retained.
}
\examples{
if (require(MEALData)){
 betavals <- betavals[1:100, ]  ## To speed up the example, the beta values are reduced
 methy <- prepareMethylationSet(betavals, pheno)
 checkSamples(methy)
}
}

