% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrNumToChar.R
\name{chrNumToChar}
\alias{chrNumToChar}
\title{Convert chr numbers to chr strings}
\usage{
chrNumToChar(vector)
}
\arguments{
\item{vector}{The vector with the chromosome numbers}
}
\value{
A vector with the chromosomes in string format.
}
\description{
Given a vector of number representing the chromosomes, convert them to string 
(e.g 1 to chr1). 23 is consider chrX, 24 is chrY, 25 is chrXY (probes shared between 
chromosomes X and Y) and 26 is chrMT.
}
\examples{
chromosomes <- c(1, 3, 4, 23, 15)
stringChrs <- chrNumToChar(chromosomes)
stringChrs
}

