% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRDAR2.R
\name{computeRDAR2}
\alias{computeRDAR2}
\title{Compute signification of RDA test}
\usage{
computeRDAR2(fullMat, varsmodel, covarsmodel = NULL, featNum, R2,
  num_permutations = 1e+05 - 1)
}
\arguments{
\item{fullMat}{Matrix with the whole genome expression or methylation values}

\item{varsmodel}{Matrix with the model}

\item{covarsmodel}{Matrix with the covariables model}

\item{featNum}{Numeric with the number of features of the RDA model}

\item{R2}{Numeric with the R2 of the RDA model}

\item{num_permutations}{Numeric with the number of permutations.}
}
\value{
Numeric vector with the probability of finding a region with the same number of probes
with a bigger R2 and the global R2.
}
\description{
Compare R2 obtained in our region of interest with the global R^2 and the R^2 of regions with 
the same number of probes.
}
