% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRanges.R
\name{createRanges}
\alias{createRanges}
\title{Create \code{GenomicRanges} from data.frame}
\usage{
createRanges()
}
\value{
Deprecated
}
\description{
Create \code{GenomicRanges} from data.frame
}
\examples{
createRanges()
}
