% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPipeline.R
\name{DAPipeline}
\alias{DAPipeline}
\title{Perform differential methylation analysis}
\usage{
DAPipeline()
}
\value{
Deprecated
}
\description{
Perform differential methylation analysis
}
\examples{
DAPipeline()
}
