% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAProbe.R
\name{DAProbe}
\alias{DAProbe}
\title{Perform per probe analysis}
\usage{
DAProbe(set, model, coefficient = 2, shrinkVar = FALSE, method = "robust",
  max_iterations = 100)
}
\arguments{
\item{set}{\code{MethylationSet}, matrix of beta values (methylation), matrix
of expression values or \code{ExpressionSet}.}

\item{model}{Matrix with the model}

\item{coefficient}{Numeric with the index of the model matrix used to perform
the analysis. If a vector is supplied, a list will be returned.}

\item{shrinkVar}{Logical indicating if shrinkange of variance should be 
performed.}

\item{method}{String indicating the method used in the regression ("ls" or 
"robust")}

\item{max_iterations}{Numeric indicating the maximum number of iterations
done in the robust method.}
}
\value{
Data.frame or list of data.frames containing intercept and slope values. 
If the set is a MethylationSet, probe's position, chromosome and the nearest
 gene are also returned.
}
\description{
Compute statistics (t estimate and p-value) for methylation or expression data
using linear or robust linear regression.
}
\examples{
if (require(minfiData)){
 mvalues <- getM(MsetEx)[1:100, ]
 model <- model.matrix(~ Sample_Group, data = pData(MsetEx)) 
 res <- DAProbe(mvalues, model, method = "ls")
 head(res)
}
}

