% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARegion.R
\name{DARegion}
\alias{DARegion}
\title{Detect regions differentially methylated}
\usage{
DARegion()
}
\value{
Deprecated
}
\description{
Detect regions differentially methylated
}
\examples{
DARegion()
}
