% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainedVariance.R
\name{explainedVariance}
\alias{explainedVariance}
\title{Calculate R2 for different variables}
\usage{
explainedVariance()
}
\description{
This function is provided for compatibility with older versions
only, and will be defunct at the next release.
}
\examples{
explainedVariance()
}
