% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{exportResults}
\alias{exportResults}
\title{Exports results data.frames to csv files.}
\usage{
exportResults(object, dir = "./", prefix = NULL,
  vars = modelVariables(object))
}
\arguments{
\item{object}{\code{MethylationResults} or \code{MethylationRegionResults}}

\item{dir}{Character with the path to export.}

\item{prefix}{Character with a prefix to be added to all file names.}

\item{vars}{Character vector with the names of the variables to be exported. Note:
names should be that of the model.}
}
\value{
Files are saved into the given folder.
}
\description{
Exports results to csv files. If more than one variable is present, subfolders
with the name of the variable are created. For each variable, four files will
be generated: probeResults.csv, dmrCateResults.csv, bumphunterResults.csv
and blockFinderResults.csv
}
\examples{
if (require(minfiData)){
set <- prepareMethylationSet(getBeta(MsetEx)[1:10,], pheno = pData(MsetEx))
methyOneVar <- DAPipeline(set, variable_names = "sex", probe_method = "ls")
exportResults(methyOneVar)
}
}

