% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterResults.R
\name{filterResults}
\alias{filterResults}
\title{Filter the data.frame obtained from probe analysis}
\usage{
filterResults(results, range, position = "position", chr = "chromosome")
}
\arguments{
\item{results}{Data.frame with the results of probe analysis}

\item{range}{\code{GenomicRanges} with the desired range.}

\item{position}{Character with the name of the column containing the positions}

\item{chr}{Character with the name of the column containing the chromosome}
}
\value{
Data.frame with the results of the probes of the range
}
\description{
Filter the data.frame obtained from probe analysis
}
