% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSet.R
\name{filterSet}
\alias{filterSet}
\title{Filter a \code{MethylationSet}, an \code{ExpressionSet} or a \code{SnpSet}}
\usage{
filterSet(set, range)
}
\arguments{
\item{set}{\code{MethylationSet}, \code{ExpressionSet} or a \code{SnpSet}}

\item{range}{\code{GenomicRanges} with the desired range.}
}
\value{
\code{MethylationSet}, \code{ExpressionSet} or a \code{SnpSet} with only 
the features of the range.
}
\description{
Filter a \code{MethylationSet}, an \code{ExpressionSet} or a \code{SnpSet}
}
\examples{
if (require(minfiData) & require(GenomicRanges)){
range <- GRanges(seqnames=Rle("chrY"), 
ranges = IRanges(3000000, end=12300000))
set <- prepareMethylationSet(MsetEx[1:100, ], data.frame(pData(MsetEx)))
set
filteredset <- filterSet(set, range)
filteredset
}
}
