% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSet.R
\name{filterSet}
\alias{filterSet}
\title{Filter a \code{MethylationSet}, an \code{ExpressionSet} or a \code{SnpSet}}
\usage{
filterSet()
}
\value{
Deprecated
}
\description{
Filter a \code{MethylationSet}, an \code{ExpressionSet} or a \code{SnpSet}
}
\examples{
filterSet()
}
