% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{getGeneVals}
\alias{getGeneVals}
\title{Get all probes related to gene}
\usage{
getGeneVals(object, gene)
}
\arguments{
\item{object}{\code{MethylationResults}}

\item{gene}{Character with the name of the gene}
}
\value{
List of data.frames with the results of the analysis of the probes belonging to 
the gene
}
\description{
Given a \code{MethylationResults} and a gene name returns the results of the 
analysis of all the probes of the gene.
}
\examples{
if (require(minfiData)){
set <- prepareMethylationSet(getBeta(MsetEx)[1:10,], pheno = pData(MsetEx))
methyOneVar <- DAPipeline(set, variable_names = "sex", probe_method = "ls")
getGeneVals(methyOneVar, "TSPY4")
}
}

