% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{getMs}
\alias{getMs}
\title{Transforms beta values to M-values}
\usage{
getMs(object, threshold = 1e-04)
}
\arguments{
\item{object}{\code{MethylationSet} or \code{AnalysisResults}}

\item{threshold}{Numeric with the threshold to avoid 0s and 1s.}
}
\value{
Matrix with the M values.
}
\description{
Given a \code{MethylationSet} or a \code{AnalysisResults} returns
the matrix of M values using a logit2 transformation. Betas equal to 0 will
be transformed to threshold and betas equal to 1, to 1 - threshold.
}
\examples{
if (require(minfiData)){
set <- prepareMethylationSet(MsetEx[1:100, ], pData(MsetEx))
mvalues <- getMs(set)
head(mvalues)
}
}

