% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultSet_getRDAresults.R
\name{getRDAresults}
\alias{getRDAresults}
\title{Get a summary of RDA results}
\usage{
getRDAresults(object)
}
\arguments{
\item{object}{\code{ResultSet}}
}
\description{
Get statistics from RDA result.
}
