% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{MEAL-defunct}
\alias{MEAL-defunct}
\title{Defunct functions}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are: multiCorrMethExprs, DAPipeline, DAProbe, DARegion, 
RDAset, filterSet, plotBestFeatures, preparePhenotype, createRanges, prepareMethylationSet
}
