% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEAL.R
\docType{package}
\name{MEAL}
\alias{MEAL}
\alias{MEAL-package}
\title{MEAL (Methylation and Expression AnaLizer): Package for analysing methylation 
and expression data}
\description{
MEAL is a package designed to facilitate the analysis methylation and expression 
data. The package can analyze one dataset and can find correlations between
methylation and expression data. MEAL has a vignette that explains the main
functionalities of the package.
}
