% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AllClasses.R, R/MethylationSet_Constructor.R, R/MethylationSet_SettersGetters.R, R/MethylationSet_checkProbes.R, R/MethylationSet_checkSamples.R
\docType{class}
\name{MethylationSet}
\alias{MethylationSet}
\alias{MethylationSet-class}
\alias{MethylationSet-methods}
\alias{betas}
\alias{betas,MethylationSet-method}
\alias{checkProbes,MethylationSet-method}
\alias{checkSamples,MethylationSet-method}
\alias{getMs,MethylationSet-method}
\alias{methylationSet}
\title{MethylationSet instances}
\usage{
methylationSet(betas, phenotypes, annotationDataFrame, annoString = "custom")

\S4method{betas}{MethylationSet}(object)

\S4method{getMs}{MethylationSet}(object, threshold = 1e-04)

\S4method{checkProbes}{MethylationSet}(object)

\S4method{checkSamples}{MethylationSet}(object)
}
\arguments{
\item{betas}{Matrix of beta values}

\item{phenotypes}{Data.frame or AnnotatedDataFrame with the phenotypes}

\item{annotationDataFrame}{Data.frame or AnnotatedDataFrame with the phenotypes with 
the annotation of the methylation sites. A column with the chromosomes named chr 
and a column with the positions names pos are required.}

\item{annoString}{Character with the name of the annotation used.}

\item{object}{\code{MethylationSet}}

\item{threshold}{Numeric with the threshold to avoid 0s and 1s.}
}
\value{
\code{MethylationSet}
}
\description{
Container with the data needed to perform methylation analysis. \code{MethylationSet}
inherits from \code{eSet} and contains \code{meth} matrix as assay data member.
}
\details{
FeatureData, which contains annotation data, is required to perform any of the
analysis.
}
\section{Methods (by generic)}{
\itemize{
\item \code{betas}: Get beta matrix

\item \code{getMs}: Get Ms values

\item \code{checkProbes}: Filter probes with annotation

\item \code{checkSamples}: Modify a \code{MethylationSet} to 
only contain common samples
}}
\section{Slots}{

\describe{
\item{\code{assayData}}{Contains matrices with equal dimensions, and with column number 
equal to nrow(phenoData). assayData must contain a matrix meth with rows representing 
features (e.g., methylation probes sets) and columns representing samples.}

\item{\code{phenoData}}{See \linkS4class{eSet}}

\item{\code{annotation}}{See \linkS4class{eSet}}

\item{\code{featureData}}{See \linkS4class{eSet}. fData should contain at least chromosome
and positions columns.}
}}
\examples{
showClass("MethylationSet")
}

