% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCorrMethExprs.R
\name{multiCorrMethExprs}
\alias{multiCorrMethExprs}
\title{Computes the correlation between methylation and expression in a genomic range}
\usage{
multiCorrMethExprs()
}
\value{
Deprecated
}
\description{
This function did not work properly and it has been deprecated. It will be made defunct in the next version.
}
\examples{
multiCorrMethExprs()
}

