% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_AllClasses.R, R/MultiDataSet-add_genexp.R, R/MultiDataSet-add_methy.R, R/MultiDataSet-add_set.R, R/MultiDataSet-add_snps.R, R/MultiDataSet-names.R, R/MultiDataSet-sampleNames.R, R/MultiDataset-subseting.R
\docType{class}
\name{MultiDataSet-class}
\alias{MultiDataSet-class}
\alias{MultiDataSet-methods}
\alias{[}
\alias{[,MultiDataSet,ANY,ANY,ANY-method}
\alias{[[,MultiDataSet,ANY,ANY-method}
\alias{add.genexp,MultiDataSet,ExpressionSet-method}
\alias{add.methy,MultiDataSet,MethylationSet-method}
\alias{add.set,MultiDataSet,eSet-method}
\alias{add.snps,MultiDataSet,SnpSet-method}
\alias{names,MultiDataSet-method}
\alias{sampleNames,MultiDataSet-method}
\title{Class \code{MultiDataSet}}
\usage{
\S4method{add.genexp}{MultiDataSet,ExpressionSet}(object, gexpSet,
  warnings = TRUE)

\S4method{add.methy}{MultiDataSet,MethylationSet}(object, methySet,
  warnings = TRUE)

\S4method{add.set}{MultiDataSet,eSet}(object, set, dataset.name,
  warnings = TRUE)

\S4method{add.snps}{MultiDataSet,SnpSet}(object, snpSet, warnings = TRUE)

\S4method{names}{MultiDataSet}(x)

\S4method{sampleNames}{MultiDataSet}(object)

\S4method{[[}{MultiDataSet,ANY,ANY}(x, i)

\S4method{[}{MultiDataSet,ANY,ANY,ANY}(x, i, j, drop = TRUE)
}
\arguments{
\item{object}{\code{MultiDataSet}}

\item{gexpSet}{\code{ExpressionSet} to be used to fill the slot.}

\item{warnings}{Logical to indicate if warnings will be displayed.}

\item{methySet}{\code{MethylationSet} to be used to fill the slot.}

\item{set}{Object derived from \code{eSet} to be used to fill the slot.}

\item{dataset.name}{Character with the name of the slot to be filled.}

\item{snpSet}{\code{SnpSet} to be used to fill the slot.}

\item{x}{\code{MultiDataSet}}

\item{i}{slot}

\item{j}{samples}

\item{drop}{Logical indicating if dropped will be applied.}
}
\value{
\code{MultiDataSet}
}
\description{
The class \code{MultiDataSet} is a superior class to store multiple
datasets in form of triplets (assayData-phenoData-featureData). The
restriction is that the samples of the multiple datasets must be the same.
}
\details{
The names of the three lists (\code{assayData}, \code{phenoData} and
\code{featureData})must be the same.
}
\section{Methods (by generic)}{
\itemize{
\item \code{add.genexp}: Method to add a slot of expression to \code{MultiDataSet}.

\item \code{add.methy}: Method to add a slot of methylation to \code{MultiDataSet}.

\item \code{add.set}: Method to add a slot to \code{MultiDataSet}.

\item \code{add.snps}: Method to add a slot of SNPs to \code{MultiDataSet}.

\item \code{names}: Get names of slots

\item \code{sampleNames}: Get sample names

\item \code{[[}: Get an eSet from a slot

\item \code{[}: Subset a MultiDataSet
}}
\section{Slots}{

\describe{
\item{\code{assayData}}{List of \code{assayData} elements.}

\item{\code{phenoData}}{List of \code{AnnotatedDataFrame} containing the phenoData
of each \code{assayData}.}

\item{\code{featureData}}{List of \code{AnnotatedDataFrame} containing the featureData
of each \code{assayData}.}

\item{\code{return_method}}{List of functions used to create the original \code{eSet} objects.}
}}

