% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalSNP.R
\name{normalSNP}
\alias{normalSNP}
\title{Normalize SNPs values}
\usage{
normalSNP(snps)
}
\description{
This function is provided for compatibility with older versions
only, and will be defunct at the next release.
}
\examples{
normalSNP()
}
