% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBestFeatures.R
\name{plotBestFeatures}
\alias{plotBestFeatures}
\title{Plot best n cpgs}
\usage{
plotBestFeatures()
}
\value{
Deprecated
}
\description{
Plot best n cpgs
}
\examples{
plotBestFeatures()
}
