% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{plotEWAS}
\alias{plotEWAS}
\title{Plot a Manhattan plot with the probe results}
\usage{
plotEWAS(object, variable = modelVariables(object)[[1]], range = NULL)
}
\arguments{
\item{object}{\code{AnalysisResults} or \code{AnalysisRegionResults}}

\item{variable}{Character with the variable name used to obtain the probe results.
Note: model name should be used. Original variable name might not be valid.}

\item{range}{GenomicRange whose cpgs will be highlighted}
}
\value{
A plot is generated on the current graphics device.
}
\description{
Plot log p-value for each chromosome positions. Highlighting cpgs inside a range
is allowed.
}
\examples{
if (require(minfiData)){
betas <- getBeta(MsetEx)[floor(seq(1, nrow(MsetEx), 10000)), ]
set <- prepareMethylationSet(betas, pheno = pData(MsetEx))
methyOneVar <- DAPipeline(set, variable_names = "sex", probe_method = "ls")
plotEWAS(methyOneVar)
}
}

