% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeature.R
\name{plotFeature}
\alias{plotFeature}
\title{Plot values of a feature}
\usage{
plotFeature(set, feat, variables = variableNames(set)[1])
}
\arguments{
\item{set}{\code{AnalysisResults}, \code{AnalysisRegionResults}, \code{ExpressionSet} or 
\code{MethylationSet}}

\item{feat}{Numeric with the index of the feature or character with its name.}

\item{variables}{Character vector with the names of the variables to be used 
in the splitting. Two variables is the maximum allowed. Note: default values 
are only valid for \code{MethylationResults} objects.}
}
\value{
A plot is generated on the current graphics device.
}
\description{
Plot values of a feature splitted by one or two variables.
}
\examples{
if (require(minfiData)){
 set <- prepareMethylationSet(getBeta(MsetEx)[1:1000, ], 
 pheno = pData(MsetEx))
 plotFeature(set, 1, variables = "Sample_Group")
 }
}

