% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLM.R
\name{plotLM}
\alias{plotLM}
\title{Plot a vector of R2}
\usage{
plotLM()
}
\description{
This function is provided for compatibility with older versions
only, and will be defunct at the next release.
}
\examples{
plotLM()
}
