% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{plotQQ}
\alias{plotQQ}
\title{QQ plot of probe analysis}
\usage{
plotQQ(object, variable = modelVariables(object)[[1]],
  main = paste("QQplot of", variable, "analysis"))
}
\arguments{
\item{object}{\code{AnalysisResults} or \code{AnalysisRegionResults}}

\item{variable}{Character with the variable name used to obtain the probe results.
Note: model name should be used. Original variable name might not be valid.}

\item{main}{Character with the plot title.}
}
\value{
A plot is generated on the current graphics device.
}
\description{
Generate a QQ plot using probe results.
}
\examples{
if (require(minfiData)){
betas <- getBeta(MsetEx)[floor(seq(1, nrow(MsetEx), 10000)), ]
set <- prepareMethylationSet(betas, pheno = data.frame(pData(MsetEx)))
methyOneVar <- DAPipeline(set, variable_names = "sex", probe_method = "ls")
plotQQ(methyOneVar)
}
}
