% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{plotRDA}
\alias{plotRDA}
\title{Plot RDA results}
\usage{
plotRDA(object, n_feat = 5, main = "RDA plot")
}
\arguments{
\item{object}{\code{AnalysisRegionResults}}

\item{n_feat}{Numeric with the number of cpgs to be highlighted.}

\item{main}{Character with the plot title.}
}
\value{
A plot is generated on the current graphics device.
}
\description{
Plot RDA results
}
\examples{
if (require(minfiData) & require(GenomicRanges)){
set <- prepareMethylationSet(getBeta(MsetEx), pheno = pData(MsetEx))
range <- GenomicRanges::GRanges(seqnames=Rle("chrY"), 
ranges = IRanges(3000000, end=12300000))
rangeNoSNPs <- DARegionAnalysis(set, variable_names = "sex", range = range)
plotRDA(rangeNoSNPs)
}
}

