% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRDAMulti.R
\name{plotRDAMulti}
\alias{plotRDAMulti}
\title{Plots a rda object from multiCorrMethExprs}
\usage{
plotRDAMulti(object, color = "black")
}
\arguments{
\item{object}{A rda object like the obtained from multiCorrMethExprs}

\item{color}{Vector with the colors of the samples. By default, all samples will be displayed in black.}
}
\value{
A plot in the graphical device
}
\description{
Function to plot the rda object obtained in \code{multiCorrMethExprs} function. Samples are displayed with a point
whose color can correspond to categorical variable. The 5 methylation probes most related to the first component and the top 
5 most related to the second component will be displayed with a blue arrow. The 5 expression probes most related 
to the first component and the top 5 most related to the second component will be labelled.
}
\details{
The legend's R-square correspond to the first two RDA components. When a factor is passed to 
color argument, the levels will also be displayed in the legend. In addition, an analysis of the variance will be
performed to evaluate if the factor can classify the samples. The p-value of this computation is also displayed
on the legend. The smaller the p-value, the better the factor can separate the samples.
}
\seealso{
\code{\link{multiCorrMethExprs}}
}

