% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{plotRegion}
\alias{plotRegion}
\title{Plot of the region}
\usage{
plotRegion(object, variable = modelVariables(object)[[1]], range = NULL,
  main = paste("Region plot of ", variable))
}
\arguments{
\item{object}{\code{AnalysisResults} or \code{AnalysisRegionResults}}

\item{variable}{Character with the variable name used to obtain the probe results.
Note: model name should be used. Original variable name might not be valid.}

\item{range}{GenomicRange whose cpgs will be shown (only for \code{AnalysisResults}
objects)}

\item{main}{Character with the plot title.}
}
\value{
A plot is generated on the current graphics device.
}
\description{
Plot of the beta values againts their position. Data is taken from probe analysis.
Cpgs with a p-value smaller than 0.05 (without adjusting) are blue and points
with a p-value greater than 0.05 are red.
}
\examples{
if (require(minfiData) & require(GenomicRanges)){
set <- prepareMethylationSet(getBeta(MsetEx), pheno = data.frame(pData(MsetEx)))
range <- GenomicRanges::GRanges(seqnames=Rle("chrY"), 
ranges = IRanges(3000000, end=12300000))
rangeNoSNPs <- DARegionAnalysis(set, variable_names = "sex", range = range)
plotRegion(rangeNoSNPs)
}
}
