% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{plotRegionR2}
\alias{plotRegionR2}
\title{Plot R2 region values}
\usage{
plotRegionR2(object, feat, ...)
}
\arguments{
\item{object}{\code{MethylationRegionResults}}

\item{feat}{Numeric with the index of the  feature or character with its name.}

\item{...}{Further arguments passed to plotLM}
}
\value{
A plot is generated on the current graphics device.
}
\description{
Plot R2 region values
}

