% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Make a Volcano plot with the probe results}
\usage{
plotVolcano(object, variable = modelVariables(object)[1], mindiff = NULL,
  main = paste("Volcano plot of", variable, "results"))
}
\arguments{
\item{object}{\code{MethylationResults} or \code{MethylationRegionResults}}

\item{variable}{Character with the variable name used to obtain the probe results.
Note: model name should be used. Original variable name might not be valid.}

\item{mindiff}{Numeric with the minimum change in methylation or expression needed to be significant}

\item{main}{Character with the plot title.}
}
\value{
A plot is generated on the current graphics device.
}
\description{
Plot log p-value versus the change in expression/methylation.
}
\examples{
if (require(minfiData)){
betas <- getBeta(MsetEx)[floor(seq(1, nrow(MsetEx), 10000)), ]
set <- prepareMethylationSet(betas, pheno = data.frame(pData(MsetEx)))
methyOneVar <- DAPipeline(set, variable_names = "sex", probe_method = "ls")
plotVolcano(methyOneVar)
}
}
