% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMethylationSet.R
\name{prepareMethylationSet}
\alias{prepareMethylationSet}
\title{Generating a \code{MethylationSet}}
\usage{
prepareMethylationSet()
}
\value{
Deprecated
}
\description{
Generating a \code{MethylationSet}
}
\examples{
prepareMethylationSet()
}
