% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePhenotype.R
\name{preparePhenotype}
\alias{preparePhenotype}
\title{Process a table of phenotypes}
\usage{
preparePhenotype()
}
\value{
Deprecated
}
\description{
Process a table of phenotypes
}
\examples{
preparePhenotype()
}
