% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDAset.R
\name{RDAset}
\alias{RDAset}
\title{Calculate RDA for a set}
\usage{
RDAset(set, varsmodel = NULL, covarsmodel = NULL)
}
\arguments{
\item{set}{\code{MethylationSet}, \code{ExpressionSet} or \code{matrix}}

\item{varsmodel}{Matrix with the model}

\item{covarsmodel}{Matrix with the covariables model}
}
\value{
Object of class \code{rda}
}
\description{
Perform RDA calculation for a \code{AnalysisRegionResults}. Feature values will
be considered the matrix X and phenotypes the matrix Y. Adjusting for covariates 
is done using a model matrix passed in covarsmodel.
}
\examples{
if (require(minfiData)){
set <- prepareMethylationSet(getBeta(MsetEx)[1:50,], pheno = data.frame(pData(MsetEx)))
model <- model.matrix(~set$age)
rda <- RDAset(set, model)
rda
}
}
\seealso{
\code{\link[vegan]{rda}}
}
