% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlockFinder.R
\name{runBlockFinder}
\alias{runBlockFinder}
\title{Run blockFinder}
\usage{
runBlockFinder(
  set,
  model,
  coefficient = 2,
  blockfinder_cutoff = 0.1,
  num_permutations = 0,
  resultSet = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{set}{\code{GenomicRatioSet}, \code{eSet} derived object or 
\code{SummarizedExperiment}}

\item{model}{Model matrix or formula to get model matrix from \code{set}.}

\item{coefficient}{Numeric with the column of model matrix used in the analysis.
(Default: 2)}

\item{blockfinder_cutoff}{Numeric with the minimum cutoff to include a probe
in a block. (Default: 0.1)}

\item{num_permutations}{Numeric with the number of permutations run to compute 
the blocks p-value. (Default: 0)}

\item{resultSet}{Should results be encapsulated in a \code{resultSet}? (Default: TRUE)}

\item{verbose}{Logical value. Should the function be verbose? (Default: FALSE)}

\item{...}{Further arguments passed to \code{blockFinder}.}
}
\value{
data.frame or \code{resultSet} with the result of \code{blockFinder}
}
\description{
Run blockFinder to a methylation dataset. This function contains all steps of
blockFinder analysis, from model.matrix creation to running the analysis.
}
\details{
runBlockFinder is a wrapper for minfi \code{blockFinder}. This function 
runs all the steps required prior running blockFinder from the methylation set
and the formula of the model. This implementation allows running blockFinder to other objects than 
\code{GenomicRatioSet}. The result can be encapsulated in a \code{ResultSet} to
take adavantege of its plotting capabilities.
}
\seealso{
\code{\link[minfi]{blockFinder}}
}
