% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{topRDAhits}
\alias{topRDAhits}
\title{Get the top features associated with the RDA}
\usage{
topRDAhits(object, pval = 0.05)
}
\arguments{
\item{object}{\code{AnalysisRegionResults}}

\item{pval}{numeric with the p-value threshold. Only features with a p-values below this threshold will be shown.}
}
\value{
data.frame with the features, the component, the correlation and the p-value
}
\description{
Get a list of the features significantly associated to the first two RDA components
}
\examples{
if (require(minfiData) & require(GenomicRanges)){
set <- prepareMethylationSet(getBeta(MsetEx), pheno = data.frame(pData(MsetEx)))
range <- GenomicRanges::GRanges(seqnames=Rle("chrY"), 
ranges = IRanges(3000000, end=12300000))
rangeNoSNPs <- DARegionAnalysis(set, variable_names = "sex", range = range)
topRDAhits(rangeNoSNPs)
}
}
