% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_AllClasses.R, R/AnalysisRegionResults_Constructor.R, R/AnalysisRegionResults_SettersGetters.R, R/AnalysisRegionResults_plotRDA.R, R/AnalysisRegionResults_plotRegionR2.R
\docType{class}
\name{AnalysisRegionResults}
\alias{AnalysisRegionResults}
\alias{AnalysisRegionResults-class}
\alias{AnalysisRegionResults-methods}
\alias{analysisRegionResults}
\alias{getRDA}
\alias{getRDA,AnalysisRegionResults-method}
\alias{getRange}
\alias{getRange,AnalysisRegionResults-method}
\alias{plotRDA,AnalysisRegionResults-method}
\alias{plotRegionR2,AnalysisRegionResults-method}
\alias{regionLM}
\alias{regionLM,AnalysisRegionResults-method}
\alias{regionPval}
\alias{regionPval,AnalysisRegionResults-method}
\alias{regionR2}
\alias{regionR2,AnalysisRegionResults-method}
\alias{snps}
\alias{snps,AnalysisRegionResults-method}
\alias{snpsPvals}
\alias{snpsPvals,AnalysisRegionResults-method}
\alias{snpsVar}
\alias{snpsVar,AnalysisRegionResults-method}
\title{AnalysisRegionResults instances}
\usage{
analysisRegionResults(analysisResults, set, range, snpspvals = data.frame(),
  regionlm = list(), relevantsnps = character(), snpsVar = as.numeric(NA),
  equation = NULL)

\S4method{getRange}{AnalysisRegionResults}(object)

\S4method{getRDA}{AnalysisRegionResults}(object)

\S4method{regionLM}{AnalysisRegionResults}(object)

\S4method{regionPval}{AnalysisRegionResults}(object)

\S4method{regionR2}{AnalysisRegionResults}(object)

\S4method{snps}{AnalysisRegionResults}(object)

\S4method{snpsPvals}{AnalysisRegionResults}(object)

\S4method{snpsVar}{AnalysisRegionResults}(object)

\S4method{plotRDA}{AnalysisRegionResults}(object, n_feat = 5)

\S4method{plotRegionR2}{AnalysisRegionResults}(object, feat, ...)
}
\arguments{
\item{analysisResults}{\code{AnalysisResults}}

\item{set}{\code{MethylationSet} or \code{ExpressionSet}}

\item{range}{\code{GenomicRanges}}

\item{snpspvals}{Data.frame obtained from \code{calculateRelevantSNPs}}

\item{regionlm}{Data.frame obtained from \code{explainedVariance}}

\item{relevantsnps}{Character vector with the relevant snps names}

\item{snpsVar}{Numeric with the variability of the SNP matrix explained by the components
used to adjust the linear model.}

\item{equation}{Character containing the formula to be used to create the model.}

\item{object}{\code{MethylationResults}}

\item{n_feat}{Numeric with the number of features to be highlighted.}

\item{feat}{Numeric with the index of the cpg or character with its name.}

\item{...}{Further arguments passed to plotLM}
}
\value{
An \code{AnalysisRegionResults}
}
\description{
\code{AnalysisResults} heir with the analyses performed in a range of the
whole genome.

AnalysisRegionResults instances
}
\section{Methods (by generic)}{
\itemize{
\item \code{getRange}: Get range where the analyses was performed

\item \code{getRDA}: Get rda object.

\item \code{regionLM}: Get R2 values of cpgs vs variables.

\item \code{regionPval}: Get p-value of lineal model R2.

\item \code{regionR2}: Get R2 of the region vs variables lineal model

\item \code{snps}: Get SNPs data

\item \code{snpsPvals}: Get p-values of correlations of snps-cpgs
pairs

\item \code{snpsVar}: Get variance of SNP matrix present in the
component used to adjusting.

\item \code{plotRDA}: Plot RDA results

\item \code{plotRegionR2}: Plot R2 region values
}}
\section{Slots}{

\describe{
\item{\code{range}}{\code{GenomicRanges} used to perform the analysis.}

\item{\code{snps}}{Character vector with the snps that are correlated to at least one cpg.}

\item{\code{snpsPvals}}{Data.frame with the results of the correlation test SNP-cpg.}

\item{\code{snpsVar}}{Numeric with the variability of the SNP matrix explained by the components
used to adjust the linear model.}

\item{\code{rda}}{\code{rda} object from \code{vegan} package with the results of RDA
analysis in the range.}

\item{\code{regionLM}}{List with the R2 of the linear model of beta values against our
variable of interest and against significant SNPs for each cpg.}

\item{\code{regionR2}}{Numeric with the R2 of the region calculated using a redundancy analysis.}

\item{\code{regionPval}}{Numeric with the pval of the region's R2.}
}}
\examples{
showClass("AnalysisRegionResults")
}

