% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRDAR2.R
\name{computeRDAR2}
\alias{computeRDAR2}
\title{Compute signification of RDA test}
\usage{
computeRDAR2(set, res, R2, nperm = 1e+06 - 1)
}
\arguments{
\item{set}{\code{MethylationSet} or \code{ExpressionSet}}

\item{res}{\code{AnalysisResults}}

\item{R2}{Numeric with the R2 obtained in the RDA}

\item{nperm}{Numeric with the number of permutations.}
}
\value{
Numeric vector with the probability of finding a region with the same number of probes
with a bigger R2 and the global R2.
}
\description{
Compare R2 obtained in our region of interest with the global R^2 and the R^2 of regions with 
the same number of probes.
}

