% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMethSNPs.R
\name{correlationMethSNPs}
\alias{correlationMethSNPs}
\title{Computes the correlation between methylation and SNPs}
\usage{
correlationMethSNPs()
}
\description{
This function is provided for compatibility with older versions
only, and will be defunct at the next release.
}
\examples{
correlationMethSNPs()
}
