% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createRanges.R
\name{createRanges}
\alias{createRanges}
\title{Create \code{GenomicRanges} from data.frame}
\usage{
createRanges(ranges)
}
\arguments{
\item{ranges}{Data.frame or matrix}
}
\value{
\code{GenomicRanges}
}
\description{
Convert a data.frame with chromosomes in the first column, starting positions in
the second one and ending position in the third one to \code{GenomicRanges}. Names
of the data.frame are preserved in the output \code{GenomicRanges}.
}
\examples{
dfranges <- data.frame(chr = c("chr1", "chr2", "chr1"), start = c(1290, 1250, 4758),
end = c(64389, 632409, 16430), stringsAsFactors = FALSE)
names(dfranges) <- c("range1", "range2", "range3")
ranges <- createRanges(dfranges)
ranges
}

