% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPipeline.R
\name{DAPipeline}
\alias{DAPipeline}
\title{Perform differential methylation analysis}
\usage{
DAPipeline(set, variable_names, variable_types = rep(NA,
  length(variable_names)), covariable_names = NULL,
  covariable_types = rep(NA, length(covariable_names)), equation = NULL,
  num_var = NULL, labels = NULL, sva = FALSE,
  region_methods = c("bumphunter", "DMRcate"), shrinkVar = FALSE,
  probe_method = "ls", max_iterations = 100, num_feat = 50,
  num_cores = 1, verbose = FALSE, ...)
}
\arguments{
\item{set}{\code{MethylationSet} or \code{ExpressionSet}}

\item{variable_names}{Character vector with the names of the variables that
will be returned as result.}

\item{variable_types}{Character vector with the types of the variables. As
default, variables type won't be changed.}

\item{covariable_names}{Character vector with the names of the variables that
will be used to adjust the model.}

\item{covariable_types}{Character vector with the types of the covariables. As
default, variables type won't be changed.}

\item{equation}{Character containing the formula to be used to create the model.}

\item{num_var}{Numeric with the number of variables in the matrix for which the
analysis will be performed. Compulsory if equation is not null.}

\item{labels}{Character vector with the labels of the variables.}

\item{sva}{Logical indicating if Surrogate Variable Analysis should be applied.}

\item{region_methods}{Character vector with the methods used in \code{DARegion}. If
"none", region analysis is not performed.}

\item{shrinkVar}{Logical indicating if shrinkage of variance should be applied 
in probe analysis.}

\item{probe_method}{Character with the type of linear regression applied in 
probe analysis ("ls" or "robust")}

\item{max_iterations}{Numeric with the maximum of iterations in the robust
regression.}

\item{num_feat}{Numeric with the minimum number of cpg beta values to be included in the 
results.}

\item{num_cores}{Numeric with the number of cores to be used.}

\item{verbose}{Logical value. If TRUE, it writes out some messages indicating progress. 
If FALSE nothing should be printed.}

\item{...}{Further arguments passsed to \code{DARegion} function.}
}
\value{
\code{MethylationResult} object
}
\description{
Wrapper for analysing differential methylation and expression at region and probe level.
}
\details{
This function is the main wrapper of the package. First, it simplifies the 
the set to only contain the common samples between phenotype and features. In addition,
it allows to change the class of the variables and to apply genomic models (more 
information on \code{preparePhenotype}). Afterwards, analysis per probe and per 
region are done merging the results in an \code{AnalysisResults} object.

Default linear model will contain a sum of the variables and covariables. If 
interactions are desired, a costum formula can be specified. In that case, variables
and covariables must also be specified in order to assure the proper work of the
resulting \code{AnalysisResult}. In addition, the number of variables of the model
for which the calculation will be done \strong{must} be specified.
}
\examples{
if (require(minfiData)){
 set <- prepareMethylationSet(matrix = getBeta(MsetEx)[1:10, ], pheno = pData(MsetEx))
 res <- DAPipeline(set, variable_names = "Sample_Group", probe_method = "ls") 
 res
}
}
\seealso{
\code{\link{preparePhenotype}}
}

