% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARegionAnalysis.R
\name{DARegionAnalysis}
\alias{DARegionAnalysis}
\title{Analyse methylation or expression in a specific range}
\usage{
DARegionAnalysis()
}
\value{
Deprecated
}
\description{
Analyse methylation or expression in a specific range
}
\examples{
DARegionAnalysis()
}
