% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEAL.R
\docType{package}
\name{MEAL}
\alias{MEAL}
\alias{MEAL-package}
\title{MEAL (Methylation and Expression AnaLizer): Package for analysing methylation and expression data}
\description{
MEAL has three different categories of important functions: processing, 
analysing and plotting.
}
\section{processing}{
 
Functions used to create \code{MEAL} objects and to modify
them. Main functions are \link{prepareMethylationSet} and \link{preparePhenotype}
}

\section{analysing}{

Functions used to perform the analysis of methylation data. \link{DAProbe} 
performs per probe analysis and \link{DARegion} performs per region analysis.
There are two wrappers: \link{DAPipeline} and \link{DARegionAnalysis} 
that performs per probe and per region analysis. The first one analyses the whole
methylation sites and the second one only a given region. Finally, 
\link{correlationMethExprs} and \link{multiCorrMethExprs} compute the correlation 
between methylation and expression probes
}

\section{plotting}{

Functions used to plot the results of the analysis. Some are interesting for
whole methylome analysis (e.g. \link{plotEWAS}) and others for analysis of one 
genomic region (e.g. \link{plotRDA})
}

