% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCorrMethExprs.R
\name{multiCorrMethExprs}
\alias{multiCorrMethExprs}
\title{Computes the correlation between methylation and expression in a genomic range}
\usage{
multiCorrMethExprs(multiset, vars_meth = NULL, vars_exprs = NULL,
  vars_meth_types = rep(NA, length(vars_meth)), vars_exprs_types = rep(NA,
  length(vars_exprs)), range = NULL)
}
\arguments{
\item{multiset}{\code{MultiDataSet} containing a \code{methylation} and an 
\code{expression} slots.}

\item{vars_meth}{Character vector with the names of the variables that will be
used to obtain the methylation residuals. By default, none is used and residuals 
are not computed.}

\item{vars_exprs}{Character vector with the names of the variables that will
be used to obtain the expression residuals. By default, none is used and
residuals are not computed.}

\item{vars_meth_types}{Character vector with the types of the methylation variables. 
By default, variables type won't be changed.}

\item{vars_exprs_types}{Character vector with the types of the expression variables. 
By default, variables type won't be changed.}

\item{range}{\code{GenomicRanges} with the desired range.}
}
\value{
An \code{rda} object
}
\description{
Estimates the correlation between methylation and expression in a range. First, 
the sets are filtered to only contain the features of the range. Then, a multivariate
approach (redundancy analysis) is applied.
}
\details{
When there are known variables that affect methylation and/or expression, their effect 
can be substracted using a linear model and then the residuals are used.
}

