% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalSNP.R
\name{normalSNP}
\alias{normalSNP}
\title{Normalize SNPs values}
\usage{
normalSNP(snps)
}
\arguments{
\item{snps}{Numerical vector or matrix representing the SNPs in the form: 0 homozygote
recessive, 1 heterozygote, 2 homozygote dominant.}
}
\value{
Numerical vector or matrix with the snps normalized.
}
\description{
SNPs values, introduced as numerical, are normalized to be used in lineal models.
}
\examples{
snps <- c(1, 0, 0, 1, 0, 0, 2, 1, 2)
normSNPs <- normalSNP(snps)
normSNPs
}
