% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBestFeatures.R
\name{plotBestFeatures}
\alias{plotBestFeatures}
\title{Plot best n cpgs}
\usage{
plotBestFeatures(set, n = 10, variables = variableNames(set)[1])
}
\arguments{
\item{set}{\code{AnalysisResults}, \code{AnalysisRegionResults}, \code{ExpressionSet} or 
\code{MethylationSet}}

\item{n}{Numeric with the number of features to be plotted.}

\item{variables}{Character vector with the names of the variables to be used 
in the splitting.}
}
\value{
Plots are created on the current graphics device.
}
\description{
Wrapper of \code{plotCPG} that plots the top n features.
}
\examples{
if (require(minfiData)){
 set <- prepareMethylationSet(getBeta(MsetEx)[1:10, ], 
 pheno = pData(MsetEx))
 plotBestFeatures(set, 2, variables = "Sample_Group")
 }
}
\seealso{
\code{\link{plotFeature}}
}

