% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDAset.R
\name{RDAset}
\alias{RDAset}
\title{Calculate RDA for a set}
\usage{
RDAset(set, equation = NULL)
}
\arguments{
\item{set}{\code{AnalysisResults}}

\item{equation}{Character with the equation used in the analysis}
}
\value{
Object of class \code{rda}
}
\description{
Perform RDA calculation for a \code{AnalysisRegionResults}. Feature values will
be considered the matrix X and phenotypes the matrix Y. Adjusting for covariates 
is done using covariable_names stored in the object.
}
\examples{
if (require(minfiData)){
set <- prepareMethylationSet(getBeta(MsetEx)[1:50,], pheno = pData(MsetEx))
methyOneVar <- DAPipeline(set, variable_names = "sex", probe_method = "ls")
rda <- RDAset(methyOneVar)
rda
}
}
\seealso{
\code{\link[vegan]{rda}}
}

